Pulled out service worker hook into its own component
